\name{ALEPlot-package}
\alias{ALEPlot-package}
\docType{package}
\title{
\packageTitle{ALEPlot}
}
\description{
\packageDescription{ALEPlot}
}
\details{
See the two individual functions \code{\link{ALEPlot}} and \code{\link{PDPlot}} that are included in this package.
}
\author{
\packageAuthor{ALEPlot}

Maintainer: \packageMaintainer{ALEPlot}
}
\references{
Apley, D. W. (2016), "Visualizing the Effects of Predictor Variables in Black Box Supervised Learning Models," submitted for publication.
}
